Fixed transaction deadlock issue
This is a fix for the transaction deadlock issue that occured with the following repro steps:
1. start transaction
2. read field
3. set value in field

The deadlock occured because the transaction grabbed a read lock at steps 2 and 3 (because
there is a read before write when setting values) before attempting to grab a write lock in
step 3. It's not possible to upgrade a read lock to a write lock, so a deadlock occured.
The fix here involves placing locking logic at the Transaction level and releasing locks
before grabbing them again to prevent situations where there are one or more readers that
are blocking the same thread from upgrading to a write lock.

UPDATE: Must also ensure that once a write lock is grabbed, it is not swapped for a read
lock, otherwise we can violate the guarantee for serializable isolation because a write
to field X followed by a read to X would allow another transaction to write to X before
the transaction has committed.