Removed reference caching of notForStorage objects
A transient issue was occuring where a BufferUnderflowException was thrown
in cases where multiple threads were trying to access the ByteBuffer of a
notForStorage object simultaneously. This usually happened when the Buffer
and the Database were both trying to read a value. In this scenario, the
Engine would spawn two different threads for the two reads. Both the Database
and the Buffer would create notForStorage comparison values, but these ended
up being the same objects because of reference caching. This was problematic
because both threads would try to copy content from the same underlying byte buffer
and the race condition would lead to the BufferUnderflowException.

The fix involves removing the reference cachin for notForStorage objects. forStorage
objects still have reference caching handled in the Byteables utility class, but this
is okay (for now) because we only expect one thread to operate on forStorage values
at a time.

There is some concern that removing reference caching could lead to gc cycles happening
more often; however, this probably won't be a huge issue and the alternative fixes to
the afeormentioned bug would either impose more guaranteed overhead (synchronizing/locking
byte buffers) or would introduce an extra level of complexity that would be prone to bugs
(i.e. byte buffer deep cloning for notForStorage object)