Return direct reference to a Record's fields instead of a unmodifiable view
There was a bug where java could not cast an UnmodifiableMap to a TreeMap which
meant that find find operations using >=, <=, >, <, and <> failed. To fix this
I've decided to return a direct reference to the Record's "fields" variable,
whose type is determined from the init() method and therefore will cast correctly.
Returning a direct reference has the potential for abuse since the variable
can be modified externally, but I don't think this will be a problem because the
variable is protected so it is only visible to subclasses who I can trust to do
the right thing.