Streamline Record model and improve write perf
This commit changes the semantics of Records. The biggest change is that
the Field abstraction is removed. Records now model data as a map of keys
to sets of values. This matches the semantics of the public api more closely.

Because of this change, the Record class must now keep track of present state
and historical indices. Some of the logic that was encapsulated in Field has
been moved up to Record.

This change is important because it does streamline the backend modeling and
makes it conform more closely to the public API, but it also allows dramatic
increases to write perf. Records now merely store a log of revisions in disk...
thats it! This means that Records will take up less space because we only ever
store one revision per write (in the past we stored a historical revision and
possibly a present state revision).

This change also means that, when writing the Record we can get away with
merely appending Revisions, which is MUCH faster than what we were doing before.
Before, we would deserialize a Record into memory, alter its state and write
the ENTIRE Record back onto disk (overwriting the previous contents). We did
this because we relied too heavily on having the in-memory representation
mirroed on disk as much as possible (to avoid having to implement B-trees).

No, we simply store a log of revisions. When the object is deserialize, we
reconstruct the in-memory indices on the fly (which is exactly what we were doing
before on deserialization). When a new write comes in, we only append that revision
to the end of the file and update the in-memory indices. We then cache the references
to Records so that we don't have to deserialize a Record that is already in memory
(because deserilization takes O(n) time no matter what)