Use bloom filters to reduce the cost of verify(key, value, record)
This change introduces the use of bloom filters when verifying the
existence of data. Previously, each call to verify() would open the
appropaite PrimaryRecord and check to see if there was presently a
mapping from key to value. Since each write makes a call to verify()
before proceeding, writes were slower because some disk activity was
involved.

With this change, we use a bloom filter for each key. Whenver data is written,
that write is added to a bloom filter for the appropriate key. Whenever
data is deserialized from disk (only from the buffer or secondary index since
these are each automatically deserialized on server startup) we add each write
or revision to the appropriate bloom filter (this means that bloom filters are
never serialized.)

When a call to verify() happens, we first check the bloom filter. If the bloom
filter says the data doesn't exist, then we return false and keep going. If the
bloom filter says the data does exist, then we open the appropriate PrimaryRecord
and check for sure.

This change has been show to speed up write perf by a factor of 5 :)