Added more stuff for new storage format:
Revision:
A wrapper for a locator, key and value that is stored in a Block

Block:
Contains Revision. This is how the Database stores data. This has lots of logic
for efficiently retrieving data

BlockFilter:
A wrapper for a bloom filter that is used in the Block

BlockIndex:
The index that records the start and end position for a locator and possibly
key in a sorted block

ByteableComposite:
A way to combine the ByteBuffers of Byteable objects

Syncable:
Can be synced to disk, but does not necessarily have to be Byteable

ByteableFunnel:
Used in a BlockFilter