CON-4: Take the revision value into account for Block sorting since we allow duplicate versions.
This commit fixes a bug where duplicate revisions appeared to be added.
The revisions had the same locator, key and version but different values.
Prior to this commit, the Block.Sorter did not take values into account when
sorting, which meant that non equal revisions actually appeared to be equal.