CON-3: Correct block indexes to give accurate find results
This commit fixes a bug that caused the find() method to return incorrect results
in some instances. The underlying cause was the fact that Block indexes were being
generated incorrectly since they used identity == instead of the equals() methods
to compare keys and locators. This commit also ensure that the results of the
find() method are in the correct sorted order