Do not use a max size for the TLock cache
Using a max size means that its possible for a lock to be evicted while
it is still being held. If that is the case, then it is possible for
a request using a Token will return a new TLock instance and we'll have
a situation where a deadlock can occur or worse, concurrent modification.
To prevent this kind of situation, we should use softValues with this
cache and only evict entries as necessary