Guard against Groovy's exposure of private fields
According to http://jira.codehaus.org/browse/GROOVY-1875, Groovy is dumb
and exposes private fields. This means that someone can get access to the
semi-sensitive client/session information stored in the "concourse" variable
or they could execute private methods of the Concourse.Client class. This
would potentially cause undefined behaviour and could be a security leak
so we must check every command to make sure that said command does not
contain a banned sequence of characters before we execute the command