CON-20: Remove uneccessary version change listener on record in AtomicOperation
This commit removes the version change listener for a modified record
in an AtomicOperation because it is unnecessary. Prior to this commit,
an AtomicOperation would fail to commit if it wrote to or read from
keyA in record and another operation wrote to keyB in record. Since keyB
and keyA are distinct, the AtomicOperation should not fail in this case.
By removing the unnecessary version change listener, the AtomicOperation
will not succeed in this case.