Fix embarassing search bug
This commit fixes a very embarassing search bug that caused some
searches to return false positive results for data in the buffer.
While iterating through Limbo to find search matches, we were not
ensuring that the key of the write matched the key of the search
request which meant that we would return any write that had a value
matching the query regardless of what the key was. This commit fixes
that issue.