Fix whitespace Text deserialization
This fix changes the deserialization logic for Text (and everything else)
such that we do not trim the resutling string that is deserialized from
the byte buffer. This fix is needed so that we can correctly handled cases
when the user stores data with leading or trailing whitespaces.

Hopefully this fix does not break anything else. In previous testing, it
seemed like it was necessary to trim strings on deserialization to prevent
flakiness, so we'll need to be on the lookout as to whether we should revert
this commit.