CON-53: Make JMX port configurable in concourse.prefs
This commit adds logic to the server start script to grep concourse.prefs
for the jmx_port paramater and copies that file to concourse.conf so that
the server can start with the user defined port for jmx connections. There
is fallback logic such that the server will start on port 9010 if there
is no jmx_port defined in concourse.prefs. This also means that the variable
can be managed by GlobalState so that we get better cross platform support
for launching the server on a windows machine :)