Improve the reliability of the lock service
This commit changes the LockService to use a Guava Cache with soft values.
The previous approach saw lots of race conditions causing IllegalMonitorStateExceptions
when lock contention was high. Now, instead of manually removing locks, we rely
on the JVM to determine if the lock if softly reachable, and if so, remove it from
the cache.