Improve the buffer transport thread efficiency
This commmit improves the efficiency of the buffer transport
thread by adding logic to avoid busy waiting if the thread has
failed to transport a Write within the last 1 second. Hopefully
this improves performance. If there is a performance degrdation,
then we should rever this commit. In actuality, the real benefit
from this commit will likely be improved CPU usage.