CON-51: "browse" implementation and refactoring
Added an implementation for browsing a key. This is in addition
to a previous commit which adds functionality to browse a record.
I also refactored the Engine to do browse directly so that the
implementation is more efficient. As a result, I also refactored
the describe functionality to use browse.