Fix issue that could occur when a write notifies an atomic operation about a version change and causes an illegal monitor state exception
Previously, when a write called onVersionChange for an atomic operation via the
Engine, the write's thread would end up calling abort() for the atomic operation and
trying to release locks held by the atomic operation. Its actually rare that the
atomic operation would be holding locks when notified about a version change, but
if it did, it would result in an illegal monitor state exception.