Implementation for lock coarsening/expansion in atomic operations
If a read within an atomic operation touches and entire record and a
subsequent write in that atomic operation touches the same record, then
we can coarsen or expand the original read lock to be a write lock and
ignore that smaller lock for the field in which the data is being written.
This should make atomic operations faster since there will be fewer locks
to grab and release. More importantly, this optimization does not sacrafice
any transaction safety