Fix ete remote object conversion bug
This commit fixes and issue in the ete test framework where objects returned
from a remote method proxy that are part of the concourse driver (i.e. Link)
can't be asserted upon or acted upon in the test logic because the object comes
from a different classpath.

To get around this, we insepct return values and perform any necessary
transformations.