CON-338: Better handling of correct maven repo
This commit changes it so that we no longer depend on correct maven
repository for upload being defined in an external .maven file. Now
the repo is programatically determined based on the pom version.

As of now, we always use the snapshot directory unless we are pushing
a version that is X.Y.Z or X.Y.Z-rcA`