GH-130: Don't return default password if one is not provided
In cases where there is interaction from the user (logging in to CaSH, authenticating
for CLI access, etc) we should not return the default password if the user supplies
a prefs file that does not specify a password. The proper thing to do is for the user
to type the password in the command prompt.

This is different for the case where a user creates a connection to Concourse or
a ConnectionPool programatically by supplying a prefs file. In that case it does make
sense for us to assume the default password because the user would not be able
to interactively supply one