GH-86: Run intializer from Concourse Server
This commit fixes issue #86 which describes a scenario where the
upgrade framework Initializer tries to run during the installation
process. This doesn't make sense because the Initializer relies on
setting values in the storage directories and the storage directories
are almost never properly configured during the installation process.

This commit fixes this issue by checking to see if the Initializer should
be run whenever the server starts. That way, we can guarantee that
the logic is only executed after the config file is properly set. And we
get the benefit of automatically running the Initializer if the storage
directories ever change.