CON-321: Updated the implementation
The previous implementation was riddled with bugs. Therefore, a complete overhaul was done and the implementation was moved to the toPostfixNotation() method for locality reasons. The resulting code is cumbersome but efficient. Additionally, the format is now insert(blob, $name = $name) rather than insert(blob, name = $name). Also, the $ on the first variable indicates the type of conjunction to use. $ for OR and $$ for AND. Any subsequent variables can use $ or $$ but will have no effect other than to identify them as variables.
The unit tests were updated to reflect this change and more were added to test the BETWEEN functionality.