CON-321: Removed support for multi-valued references
Support for multi-valued references was removed. This was done because handling multi-valued references is too complex from a code standpoint (requires look behind/ahead) and unintuitive from the users point of view (requires preprending a $ to key and distinguishing between $ (OR) and $$(AND)). Therefore, the syntax was simplified to only require a single $ on the variable. I.e. upsert(blob, "name = $name") and an error is thrown when a multi-valued reference is encountered.