Fix bug in chronologize method
After moving the chronologize method to the engine, this fixes a bug where
AtomicOperation didn't override chronologize to check the atomic state before
executing. This caused a non-repeatable read to be possible.

This commit also fixes a corner case where the user gives a future end timestamp
to the chronologize and is able to perform a non-repeatable read.