CON-590: Ignore SecurityException thrown on ManagementCli exit (#298)
This commit adds logic to ignore a SecurityException being raised in the
exit method of a ManagementCli. The exception happens because the token
that was being using in the CLI has become invalid. It is safe to ignore
the token on exit because the work that requires authorization has already
been completed.