CLOUD-1: Add access token check to invokePlugin method (#299)
* add unit tests to repro CLOUD-1

* CON-590: Ignore SecurityException thrown on ManagementCli exit

This commit adds logic to ignore a SecurityException being raised in the
exit method of a ManagementCli. The exception happens because the token
that was being using in the CLI has become invalid. It is safe to ignore
the token on exit because the work that requires authorization has already
been completed.

* add correctness check to unit test

* CLOUD-1: Add access token check to invokePlugin method

This was eroneously left off so there were some issues with plugin method
call validation.