Added generic to CommandLineInterface
This allows us to use the `options` on the superclass without having to cast it or manually call `getOptions` (which in the case of `ImportCli` returned `Options` so we had to cast it anyway.

We want to avoid unnecessary casting when the type system can validate the code.

This also doesn't necessitate a change in the implementations of this interface; you can just not pass a generic and keep using Option as it was used before.