Fix corner case with bootstrapping system version
This fixes a bug that occurs when a fresh install with upgrade tasks is pointed to an data directory with existing data at an older system version. Previously, the bootstrap process set the system version across the buffer, db and home with the latest system version (without running the upgrade tasks). This fix adjusts the bootstrap process to only set the home system version if it detects that the install is fresh