GH-426: Fix regression with upgrade task bootstrap process
Fix a bug that occurred when bootstrapping the upgrade framework for a server install with no existing data.
With the bug, the data directory versions weren't set so the upgrade frameworked assumed it was at version 0.
This caused an issue when trying to start a server version that hada minimum upgrade version requirement. The fix
involves conditionally setting the data directory versions if they are new (e.g. no data).