Tweak the lookup Strategy
Prioritize table based lookups (RECORD AND FIELD) if they are in memory before using an INDEX for
a conditonKey or orderKey. Even though the index would need to be loaded anyway, performing a lookup
using the index is slower than using the table. So, if the table is in memory it should be used because
the cache lookup to retrieve the table is the same as the cache lookup to retrieve the index