Reduce the amount of intermediate sorting that happens when going from Database -> Buffer -> Client (#563)
* cache Java format on TObject when converting from TObject to Java

When the server receives a TObject, it wraps it as a Value which is then cached on the original TObject. When Values are compared, they must be converted back to TObjects, which is just a lookup since the Value already stores its TObject representation.

However, the Convert method from TObject to Java wasn't caching the Java format on the TObject, causing multiple unnecessary conversions of the same object. This change implements caching of the Java format to improve performance.

* add clarifying note about implementation

* Do in-place filtering in Limbo/Buffer methods

* small change

* only re-sort collections in Limbo/Buffer if they're modified

* Add unit tests to verify sorting functionality

* add a BridgeSortMap implementation to efficiently take sorted Database results and incrementally sort them in the Buffer, if necessary

* For Database.browse() method, wrap the returned maps in a BridgeSortMap since the data comes from an IndexRecord and it can be assumed that those values are already sorted. This means we can avoid additional comparisons, putting the transformed data into a TreeMap

* fix formatting

* cleanup Limbo

* update changelog

* use the db comparator for case insensitive strings

* cleanup code

* Remove unnecessary re-check of key and record match in Buffer#chronologize