Add logic to detect when its better to do sorting and pagination before selecting (#565)
* Add logic to detect when its better to sort/paginate before selecting and vice versa

When selecting specific keys from records whilst paginating and sorting, it may be better to sort first and paginate (even if the sort key isn't being selected), than to select all the data first, sort and then paginate. This change adds logic to detect when that is the case. It is worth nothing that there isn't a benefit to trying to shortcut when selecting entire Records because the sort key will be gathered anyway and its not more efficient to do a one off select for ordering in this case

* update CHANGELOG

* use Command parsed orderKeys if avail

* add some unit tests

* update changelog

* add Order.isSingular check to confirm if there is only one order key

* Add frame method to Operations to faciliate sorting using index lookup when there is only one order key

* fix formatting

* fix impl bug

* fix another pre-existing bug

* force test to use traditional path

* rename variable and improve math