Improve large term index deduplication
Rename OffHeapTextSet to the more specific LargeTermIndexDeduplicator to make its purpose clearer so it isn't seen as a more general purpose utility. Also, added a brute force implementation to use as a fallback when we can calculate that there isn't enough native memory to even try to use a chronicle set. This should prevent OOMkilled on K8s where native memory isn't elastic