Improve large term index deduplication for container OS systems (#572)
* Improve large term index deduplication

Rename OffHeapTextSet to the more specific LargeTermIndexDeduplicator to make its purpose clearer so it isn't seen as a more general purpose utility. Also, added a brute force implementation to use as a fallback when we can calculate that there isn't enough native memory to even try to use a chronicle set. This should prevent OOMkilled on K8s where native memory isn't elastic

* switch estimate of capacity needed to long to prevent overflow

* add more logging as context

* fix logging and flip check to read better

* Compute more accurate sizing for memory requirements based on reality of possible substrings given length caps and realistic average key sizes

* add more logging and use off-heap memory vs native memory in log messages

* update changelog

* rename things to make more sense